BUG FIX: Dynamic E/D race condition
Basically, if the user tries to dynamically enable/disable an addon, it will not run the script since the listeners are created after the module script is loaded, which sometimes does not happen right away, like when a long project is loading. We can leave this commit in this pr since it's pretty unlikely a user will find this bug; I only found this bug as I was testing.