More dynamic addons (#4627)
* custom-block-shape

* popups

* full-signature

* edtior-messages

* exact-count

I felt that there was no need to remove/revert the old counts since it's a waste of time and not useful.

* mute-project

* BUG FIX: Dynamic E/D race condition

Basically, if the user tries to dynamically enable/disable an addon, it will not run the script since the listeners are created after the module script is loaded, which sometimes does not happen right away, like when a long project is loading. We can leave this commit in this pr since it's pretty unlikely a user will find this bug; I only found this bug as I was testing.

* progress-bar

* better-img-uploads

* initialise-sprite-position

* load-extensions

* longer-wiwo

* project-info

* block-count

* Remove dynamicDisable from exact-count

* editor-devtools

I didn't add dynamicDisable to this addon since the addon will, in the future, be split up into many different addons. (#3700)

* data-category-tweaks-v2

* feature-unshared

* blocks2image

* last-edit-tooltip

* blocks2image: return items earlier

* true-youtube-links: dynamic enable only

* forum-search

* editor-sounds

* ctrl-enter-post

* disable-auto-save