Fix bug when another language as input before text
This bug was actually in devtools before I made this PR. To reproduce, set the editor language to "Magyar", go to the devtools search and find a "when i receive block", click on it and notice it says "0". This happens with other languages too, when the block's input is on the left side instead of the right. It has been fixed in this commit.