Remove worst case O(n^2) time complexity
This fixes a performance regression when editing projects with deeply
nested scripts. The problem with the old logic is that when you edit
something deep into a script, that block and every parent will each each
re-stripe everything from the top down, which is unnecessary and has
O(n^2) time complexity (although it isn't noticeable issues until blocks
are nested >100 levels deep).
The fix is to re-stripe everything exactly once, which is now the
responsibility of the "top" block.
As a side-effect, I think the code is more straightforward and no longer
mutates block objects (except the render trap)