Extract find bar and jump to definition from devtools to separate addons (#4817)
* Create find bar addon

* Format code

* Fix class name

* Make navigation arrays global

* Remove help button and locales

* Remove unused styles/code

* Fix darkmode styling

* Fix lgtm errors

* Clean up findbar userscript file a bit

* Implement dynamic enable and disable

* Update addon.json

* Organize find bar js more.

TODO: Add more classes?

* Allow searching for sounds

* Fix bug where searching is delayed

* Organize find bar js code more

hey i created a new class :)

* Remove commented code

* Create Jump to definition addon

* Fix jump-to-def addon.json

* Merge multi into dropdown class

* Final find bar cleanup

Although I may possibly add a new carousel class ;P

* Remove unused locales

* Create carousel class

* Fix lgtm error

* Change names, descriptions and infos.

* Fix bug when another language as input before text

This bug was actually in devtools before I made this PR. To reproduce, set the editor language to "Magyar", go to the devtools search and find a "when i receive block", click on it and notice it says "0". This happens with other languages too, when the block's input is on the left side instead of the right. It has been fixed in this commit.

* Pollute Blockly.Gesture to get block middle click

* Add transition code

Read comment for what we are doing in this transition

* Leave find bar empty if selected tab is not known

* Create find bar class once and pollute blockly once

* Remove unneeded code

We already listen for when the input looses focus, which can happen when the mouse is clicked.

* Remove unused devtools functions

* Fix migration code

* Apply suggestions from code review

* Fix conflicts with editor-theme3

later in the future we should probably organize the css a bit better and maybe make the js code for this part a little neater too.

Co-authored-by: TheColaber <TheColaber@users.noreply.github.com>
Co-authored-by: apple502j <33279053+apple502j@users.noreply.github.com>