2 small bug fixes on search pages
(1) Infinite Scrolling inserts a synthetic "Page n" cell that spans the full row, so has no sibling. On search pages, we check sibling cells for the topic's forum. Skip this synthetic row when gathering.

(2) query for tableBodyNode (tbody) early, and if not available and not on mobile, we won't be able to process this page.

It happens on search pages with a "&show_as=posts" query parameter

https://scratch.mit.edu/discuss/search/?action=show_user&show_as=posts

Error didn't affect users, since the page shouldn't be highlighted anyway. But this keeps the error log clean.

We could also avoid running on pages with this query parameter if someone wants to suggest a fancier regex for that. But this early check for tbody is still reasonable in case there are any other search pages slipping through the cracks.