Improve scrollbars
Makes scrollbars work better with dark mode. Adds 1px of space between the scrollbar and the edge of the element because rounded scrollbars look weird if they touch the edge.

According to the documentation, settings.if evaluates as settings OR addonEnabled. For some reason, the condition for stage_compact_scrollbars.css evaluates as AND, not OR, which is actually exactly what I need but probably not the intended behavior.