New addon: compact editor (#4018)
* Initial commit (WIP)

* Fix block area height on enable/disable

* Compact sound editor, paint editor consistency

* Format code

* Tutorial and debugger modal support

* Update addon.json

* Decrease size of box shadows

* Format code

* Update addon.json

* Compactify project player toolbar

* Format code

* Compatibility with `sprite-properties` addon

* Resize tool size icon

* Misc. fixes and touch-ups

* Small changes to addon.json

* Keep blue highlight on selected sprite

This makes it easier to see which sprite is selected. Removing the blue only makes sense for costume and sound tiles because the addon reduces their padding.

* Blue delete button

* Larger category bubbles and icons

This makes icons 4px larger and reduces padding by 2px on each side, so the result is equally compact.

* Don't make changes to the project page

* Improve scrollbars

Makes scrollbars work better with dark mode. Adds 1px of space between the scrollbar and the edge of the element because rounded scrollbars look weird if they touch the edge.

According to the documentation, settings.if evaluates as settings OR addonEnabled. For some reason, the condition for stage_compact_scrollbars.css evaluates as AND, not OR, which is actually exactly what I need but probably not the intended behavior.

* Compatibility with Scratch update

* Don't change flyout width

This would only make sense if blocks were also made smaller by the same addon.

* Improve compatibility with columns

Uses class instead of attribute to make the CSS selectors simpler

* Compatibility with onion-skinning and paint-snap

* Fix Convert to Bitmap button

* Compatibility with hide-stage

* Compatibility with sprite-properties

* Make hiding labels optional

* Add tooltips if labels are hidden

* Detect menu bar resizes after switching to project page and back

* Update versionAdded

* Fix RTL bugs

* Add reduxEvents to waitForElement(sound-editor)

* Add reduxEvents and reduxCondition to waitForElement(menuBar)

* Fix columns compatibility after changing language

* Do not affect project player fullscreen

* "Code area" instead of workspace

* Add featured tag

---------

Co-authored-by: lisa-wolfgang <lisa-wolfgang@users.noreply.github.com>
Co-authored-by: Maximouse <51849865+mxmou@users.noreply.github.com>
Co-authored-by: World_Languages <mundofinky@gmail.com>