Allow more inputs in editor-number-arrow-keys <!-- Which issue(s) does this pull request fix or resolve? If there aren't any, please submit one first unless this is a hotfix or minor string update. -->
Separates the support for other inputs into its own pr. Keep the changes of how it scans for type=number and type=text for simplicity.

### Changes

Allows this addon to work in the colour picker inputs, onion skinning inputs, and paint snapping inputs.

### Reason for changes

Every input should be accepted.

### Tests

Tested in brave while it was a part of the math-in-inputs pr